create the ui
