+++ resources/views/panel/scheduledConference/resources/registrant-resource/pages/participant-attendance.blade.php
@@ -1,3 +1,3 @@
 <x-filament-panels::page>
-    {{ $this->table }}
+    {{ $this->infolist }}
 </x-filament-panels::page>
